+++ jmh-core/src/main/java/org/openjdk/jmh/profile/GCProfiler.java
@@ -98,7 +98,7 @@ public class GCProfiler implements InternalProfiler {
                                     Double.NaN,
                             "MB/sec", AggregationPolicy.AVG));
             if (allocated != 0) {
-                long allOps = iResult.getMetadata().getAllOps();
+                double allOps = iResult.getMetadata().getAllOps();
                 results.add(new ScalarResult(Defaults.PREFIX + "gc.alloc.rate.norm",
                                 (allOps != 0) ?
                                         1.0 * allocated / allOps :
